requirements.txt
torch==1.8.1
torchvision==0.9.1
numpy
pandas
matplotlib
seaborne
acelerate
timm
pillow
